Changed definition of paths so that the codomain is Set instead of Type.
If we specify Type as the codomain of paths, Coq will actually infer that
paths maps into Prop (which is one of the possible Type's...) This is very
bad for HoTT. So it is better to specify that the codomain is Set. I am
not sure what this implies in terms of universes and sizes. It seems
strange that, for example, the identity type of a universe U should map
into Set. Surely, U ~~> U ought to live at the same level as U? Nevertheless,
this seems the best we can do with the current version of Coq.