An improved treatment of function extensionality.
We optimize proofs in Funext.v to make them conceptually cleaner.
In particular, from the original, rather monolithic proof, we extract
an obvservation about the axiom of choice being an equivalence from
the space of totality of a relation to the space of choice functions
of a relation. It is likely that this can be further improved, i.e.,
the extra condition about contractibility in the proof of ac_equiv
might be better expressed in terms of is_prop.