Further improvements to the homotopic axiom of choice.
Going through FunextEquivalences, I noticed that section_total_equiv
is nothing but the axiom of choice! Except, in comparison with ac_equiv
(which is now ac_equiv') it uses strong extensionality and no assumption
on the relation. So, I improved Funext.v to clarify the situation with
the axiom of choice, which is quite interesting (read the comments).