Moved basic definitions to Overture.v to simplify dependencies between files.
Files became dependent on each other in circular ways, so we had to place theorems
in unnatural places to remove the circularities. Therefore, I moved basic concept
definitions into Overture.v. For example, the definition of hlevel is in Overture.v,
but all general results are in HLevel.v. Thus, a result about equivalences is allowed
to mention contractibility, and vice versa.

To shorten Overture.v I moved the lengthy but boring groupid structure of path spaces
to PathGroupoids.v.