Change the level of the infix == notation
This way it's more compatible with the standard library, e.g.,
theories/Classes/SetoidClass.v.

Also, put it in [type_scope], because it's only useful when Coq is
expecting something of type [Type], and because scoped notations are
nicer to deal with when building on top of libraries.  We might instead
want to put it in its own scope, like how there is a different variant
in [Q_scope] in the standard library, but [type_scope] is a reasonable
default.