Add .travis.yml for a travis-ci
This file will, if Travis CI is enabled for the repo by going to
https://travis-ci.org/profile/, add an automated build to check pull
requests for compatibility with HoTT/coq stable.  It is simple to modify
it to use trunk instead, should we desire.  Pull requests will be marked
with a green circle if they build, a yellow circle while they are
building, or a red circle if they fail, approximately 15-30 minutes
after they are submitted.