Use coq_makefile for the HoTT library
This way, we support `make html` and `make TIMED=1` and other niceties
that coq_makefile gives us.  Unfortunately, I haven't figured out how to
suppress the warning message about overwriting the `clean` target.  I'm
pretty sure I got the ordering right, though, so that the `clean` target
does what we want it to.  (There's also the slightly wonky dependency
that `clean` depends on `Makefile.coq`, but then goes and deletes it.
This is made worse by the fact that I had to put the stdlib .vo files as
a dependency for `Makefile.coq`, because I haven't figured out any other
way to inject them as dependencies in the targets of `Makefile.coq`.)