Add --without-bundled-coq to autogen, fix make
You can now call './autogen.sh --without-bundled-coq' to not init the
submodule.

Passing COQBIN to ./configure now disables the execution of ./configure
in coq-HoTT.

When not using the bundled Coq, we no longer make the miniopt target in
coq-HoTT.  I was originally doing this to build just the required
plugins, so we can statically link to them.  But I gave up on this
eventually (at least for now), and so we no longer make the miniopt
target.