Remove trailing slashes from the `ln` invocation
While it worked fine with GNU tools, it doesn't work correctly with BSD
tools, which is what Mac OSX uses.  This closes #201 (again).
Conceptually, the reason to not have the trailing slashes is so that,
e.g., `coq/dev` resolves to `coq-HoTT/dev`, and `coq/dev/` resolves to
`coq-HoTT/dev/`, rather than having `coq/dev` resolve to `coq-HoTT/dev/`
and `coq/dev/` resolve to `coq-HoTT/dev//`.