Add separate targets for hott-{categories,core}
This closes #257.

This closes #240.

You now need to update Makefile_targets.mk if you add a file.  This will
go in some documentation soon; the wiki and/or some files.

I'm not sure what the best order of files is in Makefile_targets.mk
(which determines the order in the TOC).  In particular, I'm not sure if
we want to enforce a strict "dependencies come first" order, or if we
want a conceptual grouping, and put things in the same folder together.

Perhaps we should also have a "strict" target of the makefile that fails
if there are any commited .v files which are not listed in
Makefile_targets.mk?

Currently, we, by default, build both hott-categories and hott-core.
We may decide to change this, and, if we do, we should update travis to
build both.  Perhaps we should also add a target for the HoTT book in
contrib?