Remove support for ssreflect
The ssreflect support has been inactive for a long time now. It would
take some effort to make it work, and it is unlikely this will happen.
So I am ripping it out. If in the future anyone wants to put it back
in, they can have a look at the history and go from there.