Reorganize categories
It seems better to have absolute names be, e.g., [HoTT.categories]
rather than [HoTT.categories.categories].  This has the questionable
advantage of allowing (actually, requiring, a la
https://coq.inria.fr/bugs/show_bug.cgi?id=3676) us to not manually
recreate subdirectory structure in [Module]s in the interface files.
The downside is that we don't get automatic renaming of the
subdirectories if we [Include] the interface modules elsewhere.

This closes #503.