Actually install all of the symlinks to stdlib
Previously, we had separate logic in Makefile.am and configure.ac for
dealing with symlinks to the standard library.  Since no one was testing
the install target regularly, they had gotten out of sync.  This was
causing hoqide to fail when the library was installed.

Now, the configure script makes a list of symlinks for the makefile to
copy over in the install target.  This should fix hoqide.  Additionally,
the install target should now work fine on (Windows) configurations
where symlinks don't work.

Finally, we now test each stdlib directory for existence before
symlinking to it.  This prevents us from installing dangling symlinks to
the system, which wouldn't serve any purpose.