Add some notations for equivalences
Also, correct equiv_scope.  The new rule for [equiv_scope] is that it
should match [function_scope], [morphism_scope], [type_scope]:
[equiv_scope] is for things that build [Equiv]s (just like [type_scope]
is for things that build [Type]s).

This permits and necessitates removing many [Local Open Scope
equiv_scope]; things that expect [Equiv] now get the notations
automatically, and since [<~>] is in [type_scope] and [^-1] is in
[function_scope], we don't need [equiv_scope] for those.  We also must
not have [equiv_scope] on top of [function_scope], or else we'll pick up
the wrong [o].