Fix [rewrite@A]
When the hypothesis contained [a @ (b @ c)] where the implicit arguments
to [concat] did not match the implicit arguments inferred for [@concat _
_ _ _ a (@concat _ _ _ _ b c)] (according to [pattern]),
[left_associate_concat_in] looped, repeatedly finding [a @ (b @ c)], and
failing to associate it to the left, introducing constant transports
that did nothing.

This is now fixed, by picking up and passing the implicit arguments,
explicitly.

The comment in Modalities/ReflectiveSubuniverse.v has been updated; a
naive use of [rewrite@A] causes the subsequent rewrite to fail.  I
haven't tracked it down, but I suspect @mikeshulman wouldn't have much
trouble updating the file.

This fixes #744; the test is now added to Tests.v.