No substitution of "Book_X_Y" at all
I changed my mind: Exercise 2.3 was still broken, while it made sense to call the definition there Book_2_1_concatenation4. We probably don't need the substitution anymore, as the numbering in book is not likely to change. If this assumption proves to be false, the substitution could be reintroduced on an ad-hoc basis.