Set Coq version to 8.8.1
This commit is intended to be the last commit that we guarantee builds
with Coq v8.8.  This commit (or the merge commit of this PR) will be the
target of the V8.8 tag, and after this commit is merged, we will drop
support for Coq 8.8 in preparation to merge the PR that allows
compatibility with numeral notations.