Declare argument scopes explicitly for `AdjunctionUnitCounit`
HoTT was relying on a Coq bug which made some bound scopes visible even
if they were not imported in some scenarios. We avoid this by declaring explicitly some argument scopes.

This is in preparation for https://github.com/coq/coq/pull/10181, but
should be backward compatible (remains to be tested).