Allow compilation in multiple git working trees
The `git worktree` command allows to have several working trees for a
single repository.  The additional working trees then get a _file_
`.git` containing the gitdir, rather than a full `.git` directory.  In
order to support this, the build files should additionally allow
`.git` to be a file rather than a directory.