Bump coq-dpdgraph coq version to master
The tip of `coq-master` is too new to work with 8.13, so we build
coq-dpdgraph against master.  This means that random changes in Coq
master, even with `coq-dpdgraph` overlays merged, may break our dpdgraph
build.  So we should ideally switch this back to a non-master version of
Coq once 8.14 is released.

For now, though, if we want to track the main dpdgraph repo and support
a new enough OCaml, we need this version.