Adjoint functors
We define a notion of adjoint functors using natural equivalences of
Hom-types. We prove that this is equivalent to a unit-counit
characterization of adjoints under strongness assumptions for the
categories involved. We also prove that postcomposition applied to
adjoint functors form an adjoint pair as well. We also have lemmas
about composition of adjoint functors.