Weaken 'Decidable' to 'Stable' in hypotheses.
We could still leave the original theorem, named "merely_inhabited_iff_inhabited_decidable", or we could leave it up to typeclass inference to patch that gap, as there is a Decidable -> Stable hint. Not clear which is better.

New version of the theorem lets us conclude, for example,  merely (~A) -> ~A for any A, even if A is not decidable.