wildcat: binary products
We define what it means for a wild cat to have binary products and then
show Type, pType and Group to have binary products. This notion of
product is only 1-coherent as it will not be the correct notion of a
product in a bicategory. But this doesn't matter for now as we haven't
even defined a full bicategory yet.

AbGroup will be left out for now as a notion of biproduct will follow
soon.

Signed-off-by: Ali Caglayan <alizter@gmail.com>