add universe constraints in GraphQuotient
GraphQuotient takes 3 universe levels. The first is the level of the
type we are quotienting. The second is the level of the relation we are
quotienting by. Finally, the third is the level the quotient type will
live in.

Before we didn't enforce that the first and second universe levels are
less than or equal to the third. We add those constraints in.

Signed-off-by: Ali Caglayan <alizter@gmail.com>