Nat: remove unused hints
Hint databases require much more care than afforded here. The current
situation seems to be adding every lemma etc as a hint. This will have
to be rethought if we want a auto solver for nat lemmas, but in the mean
while we should avoid using auto and rely on typeclasses instead.

Signed-off-by: Ali Caglayan <alizter@gmail.com>

<!-- ps-id: 712ec5e9-fad1-42b7-8141-b446ab69b604 -->