cleanup: remove do_nat tactic
This was a strange tactic that acted like `do n ltac_term` but `n` was
allowed to be an internal `nat`. We don't particularly have any proofs
that use internal `nat`s this way, so it's difficult to see how this
would be useful. If a real usecase arises in the future we can always
add it back in, otherwise it doesn't seem to serve a purpose in Basics/.

Signed-off-by: Ali Caglayan <alizter@gmail.com>

<!-- ps-id: ae9c7bf6-30a3-493c-a984-900e8be87392 -->