cleanup: remove Utf8.v
It was unused and probably not the best practice. We use UTF8 sparingly
throughout the library and as long as we don't over do it, it seems to
be fine. I don't believe anybody external is relying on this file, but
if they are we can always add it back in on request.

Signed-off-by: Ali Caglayan <alizter@gmail.com>

<!-- ps-id: 58dd3673-5eca-4b5b-96c5-f990a72f0bc4 -->