mathclasses improvements
We introduce a whole set of improvements to the mathclasses library. The
main one being that (^) is now the notation of group inverses. Doing so
required lots of tweaking of hints throughout the library and I took the
time to fix some obviously wrong things as I encountered them, including
rewriting various proofs.

Signed-off-by: Ali Caglayan <alizter@gmail.com>