Be more selective with WildCat and other imports
Before moving SetoidRewriting to WildCat folder, we make sure other
files in the library don't Require the WildCat.v meta-file, so they
won't indirectly get SetoidRewriting imported.  This will make the
library continue to build with Coq 8.20, and will also reduce
unnecessary rebuilds.

We also reduce the dependence on Pointed/ and Algebra/, as some of
those files will indirectly depend on SetoidRewriting in a future
commit.